improve code locality - combine sysctl sources
Combine all non-cgo sysctl based data sources into one file
to improve the code locality.
