+++ providers/freebsd/defs_freebsd.go
@@ -28,8 +28,6 @@ package freebsd
 */
 import "C"
 
-type vmTotal C.struct_vmtotal
-
 type kvmSwap C.struct_kvm_swap
 
 type clockInfo C.struct_clockinfo
