+++ providers/freebsd/memory_freebsd_cgo.go
@@ -38,92 +38,6 @@ import (
 	"golang.org/x/sys/unix"
 )
 
-func totalPhysicalMem() (uint64, error) {
-	const mib = "hw.physmem"
-
-	v, err := unix.SysctlUint64(mib)
-	if err != nil {
-		return 0, fmt.Errorf("failed to get %s: %w", mib, err)
-	}
-	return v, nil
-}
-
-func pageSizeBytes() (uint32, error) {
-	const mib = "vm.stats.vm.v_page_size"
-
-	v, err := unix.SysctlUint32(mib)
-	if err != nil {
-		return 0, fmt.Errorf("failed to get %s: %w", mib, err)
-	}
-
-	return v, nil
-}
-
-func activePageCount() (uint32, error) {
-	const mib = "vm.stats.vm.v_active_count"
-
-	v, err := unix.SysctlUint32(mib)
-	if err != nil {
-		return 0, fmt.Errorf("failed to get %s: %w", mib, err)
-	}
-	return v, nil
-}
-
-func wirePageCount() (uint32, error) {
-	const mib = "vm.stats.vm.v_wire_count"
-
-	v, err := unix.SysctlUint32(mib)
-	if err != nil {
-		return 0, fmt.Errorf("failed to get %s: %w", mib, err)
-	}
-	return v, nil
-}
-
-func inactivePageCount() (uint32, error) {
-	const mib = "vm.stats.vm.v_inactive_count"
-
-	v, err := unix.SysctlUint32(mib)
-	if err != nil {
-		return 0, fmt.Errorf("failed to get %s: %w", mib, err)
-	}
-
-	return v, nil
-}
-
-func cachePageCount() (uint32, error) {
-	const mib = "vm.stats.vm.v_cache_count"
-
-	v, err := unix.SysctlUint32(mib)
-	if err != nil {
-		return 0, fmt.Errorf("failed to get %s: %w", mib, err)
-	}
-
-	return v, nil
-}
-
-func freePageCount() (uint32, error) {
-	const mib = "vm.stats.vm.v_free_count"
-
-	v, err := unix.SysctlUint32(mib)
-	if err != nil {
-		return 0, fmt.Errorf("failed to get %s: %w", mib, err)
-	}
-
-	return v, nil
-}
-
-// buffersUsedBytes returns the number memory bytes used as disk cache.
-func buffersUsedBytes() (uint64, error) {
-	const mib = "vfs.bufspace"
-
-	v, err := unix.SysctlUint64(mib)
-	if err != nil {
-		return 0, fmt.Errorf("failed to get %s: %w", mib, err)
-	}
-
-	return v, nil
-}
-
 func kvmGetSwapInfo() (*kvmSwap, error) {
 	var errstr *C.char
 	kd := C.kvm_open(nil, nil, nil, unix.O_RDONLY, errstr)
